Move all app access related logic to a service
